Tag name modified (#4)
This pull request includes a minor update to the `README.md` file for
the Dev Container for Ansible. The change involves modifying the base
image tag to be more flexible.

*
[`README.md`](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5L9-R9):
Changed the base image tag from `python:latest` to `python:*tag*` to
allow for more flexibility in specifying the Python version.